@@ -8,12 +8,12 @@
     "preview": "vite preview"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.4",
     "@tsconfig/svelte": "^5.0.0",
     "@types/node": "^20.0.0",
     "moment": "^2.30.1",
     "sass": "^1.90.0",
-    "svelte": "^4.0.0",
+    "svelte": "^5.38.0",
     "svelte-preprocess": "^5.1.0",
     "typescript": "^5.0.0",
     "vite": "^5.0.0"
